requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
Jinja2
python-multipart
pydantic
pydantic-settings
pytest
httpx
pytest-asyncio
pydantic[email]
email-validator
black
isort
